updated js model
