+++ public/assets/landing_page.png
