+++ public/js/auth/index.js
