+++ public/js/gwas.js
@@ -1,0 +1,13 @@
+
+export {
+    makeChrPlot,
+    makeLocusPlot,
+    initAll
+} from "./gwas/plots.js";
+
+export {
+    queryDataSourcesFromSelectors,
+    createListeners
+} from "./gwas/services.js";
+
+export { PlabGenes } from "./gwas/genes.js";
