+++ public/templates/gwas.html
@@ -7,12 +7,6 @@
     src="https://cdn.jsdelivr.net/npm/d3@^5.16.0"></script>
 <script type="text/javascript"
     src="https://cdn.jsdelivr.net/npm/locuszoom@0.14.0/dist/locuszoom.app.min.js"></script>
-<script type="text/javascript"
-      src="/public/static/js/gwas.js"></script>
-<script type="text/javascript"
-    src="/public/static/js/genes.js"></script>
-<script type="text/javascript"
-      src="/public/static/js/locuszoom_gwas.js"></script>
 <link rel="stylesheet"
       type="text/css"
       href="/public/static/css/site.css">
@@ -67,7 +61,11 @@
     </div>
 </body>
 
-<script type="text/javascript">
+<script type="module" src="/public/js/gwas.js">
+
+    console.log(LocusZoom)
+    console.log(PlabGenes)
+    LocusZoom.DataLayers.add("plab_genes", PlabGenes)
     let plotListener = null
     let plots = []
 
@@ -76,88 +74,26 @@
     queryElements.set("phenotype", document.getElementById("phenotype"));
     queryElements.set("chr", document.getElementById("chr"));
 
-    queryElements.get("projectId").eventProcessor = getPhenotypeFromQuery("/api/gwas/phenotypes",
-                                                            "projectId");
-    queryElements.get("phenotype").eventProcessor = getChrFromQuery("/api/gwas/chr",
-                                                            "projectId",
-                                                            "phenotype");
+    // TODO: How do I reset all options under page refresh
+    queryElements.get("projectId").eventProcessor = queryDataSourcesFromSelectors("/api/gwas/phenotypes",
+        [],
+        "projectId",
+        ["phenotype", "chr"]);
+    queryElements.get("phenotype").eventProcessor = queryDataSourcesFromSelectors("/api/gwas/chr",
+        ["projectId"],
+        "phenotype",
+        ["chr"]);
+
     queryElements.get("chr").eventProcessor = initializeGwasPlots("projectId",
-                                                            "phenotype",
-                                                            "chr",
-                                                            {
-                                                                chrOverview: "lzChrOverview",
-                                                                locusOfInterest: "lzLocusOfInterest"
-                                                            });
+        "phenotype",
+        "chr",
+        {
+            chrOverview: "lzChrOverview",
+            locusOfInterest: "lzLocusOfInterest"
+        });
 
     createListeners(queryElements)
 </script>
 
 
 </html>
-
-<!-- 
-                        {
-                            namespace:{'assoc':'myAssocTest'},
-                            id:"assoc",
-                            type: "scatter",
-                            x_axis: {field: 'assoc:position'},
-                            y_axis: {
-                                axis: 1,
-                                field: 'assoc:log_pvalue',
-                                floor: 0,
-                                upper_buffer: 0.10,
-                                min_extent: [0, 10]
-                            }
-                        }
-                    ]
-            //responsive_resize: true,
-            //min_region_scale: 20000,
-            //max_region_scale: 1000000,
-    class MyAdapter extends LocusZoom.Adapters.get("BaseLZAdapter") {
-        _normalizeResponse(response_text, options) {
-            let data = super._normalizeResponse(...arguments);
-            // Most portaldev endpoints (though not all) store the desired response in just one specific part of the payload
-            data = data.data || data;
-
-            if (Array.isArray(data)) {
-                // Already in the desired form
-                return data;
-            }
-            console.error("Not working");
-            throw new Error("weird");
-
-        }
-    }
-    class MyGeneAdapter extends LocusZoom.Adapters.get("BaseLZAdapter") {};
-
-    assoc_adapter = new MyAdapter({url: "/public/ds/assoc_corrected.json"});
-    recomb_adapter = new MyAdapter({url: "/public/ds/recomb_corrected.json"});
-    ld_adapter = new MyAdapter({url: "/public/ds/ld_corrected.json"});
-    gene_adapter = new MyAdapter({url: "/public/ds/genes.json"});
-    constraint_adapter = new MyAdapter({url: "/public/ds/constraint.json"});
-
-    let data_sources = new LocusZoom.DataSources()
-        .add("assoc",   assoc_adapter)
-        .add("recomb",  recomb_adapter)
-        .add("ld",      ld_adapter)
-        .add("gene",    ["GeneLZ", {url: "/public/ds/genes.json"}])
-        .add("constraint", constraint_adapter);
-
-    let layout = {
-        width: 800,
-        state: {
-            genome_build: "GRCh37",  // or whatever build your data uses
-            chr: 10,                 // specify your chromosome
-            start: 114550452,        // start position
-            end: 115067678          // end position
-        },
-        panels: [
-            LocusZoom.Layouts.get("panel", "association"),
-            LocusZoom.Layouts.get("panel", "genes")
-        ]
-    };
-
-
-
-    const plot = LocusZoom.populate("#lz-plot", data_sources, layout);
-        -->
